+++ addon/components/power-select/options.js
@@ -12,18 +12,13 @@ export default Ember.Component.extend({
   didInsertElement() {
     this._super(...arguments);
     if (this.get('role') === 'group') { return; }
-
-    this.element.addEventListener('mouseup', e => {
-      const optionItem = Ember.$(e.target).closest('[data-option-index]');
-      if (!optionItem) { return; }
-      this.get('select.actions.choose')(this._optionFromIndex(optionItem[0].dataset.optionIndex), e);
-    });
-    this.element.addEventListener('mouseover', e => {
-      if (e.target.dataset.optionIndex) {
-        this.get('select.actions.highlight')(this._optionFromIndex(e.target.dataset.optionIndex), e);
-      }
-    });
-
+    let findOptionAndPerform = (action, e) => {
+      let optionItem = Ember.$(e.target).closest('[data-option-index]');
+      if (!optionItem || !(0 in optionItem)) { return; }
+      action(this._optionFromIndex(optionItem[0].dataset.optionIndex), e);
+    };
+    this.element.addEventListener('mouseup', e => findOptionAndPerform(this.get('select.actions.choose'), e));
+    this.element.addEventListener('mouseover', e => findOptionAndPerform(this.get('select.actions.highlight'), e));
     if (this.get('isTouchDevice')) {
       this._addTouchEvents();
     }
@@ -55,7 +50,7 @@ export default Ember.Component.extend({
     let parts = index.split('.');
     let options = this.get('options');
     if (!options.objectAt) {
-      option = Ember.A(options);
+      options = Ember.A(options);
     }
     let option = options.objectAt(parseInt(parts[0], 10));
     for (let i = 1; i < parts.length; i++) {