+++ test-support/helpers/ember-power-select.js
@@ -10,8 +10,8 @@ function typeText(selector, text) {
   $selector[0].dispatchEvent(event)
 }
 
-export function nativeMouseDown(selectorOrDomElement, options = {}) {
-  let event = new window.Event('mousedown', { bubbles: true, cancelable: true, view: window });
+function fireNativeMouseEvent(eventType, selectorOrDomElement, options = {}) {
+  let event = new window.Event(eventType, { bubbles: true, cancelable: true, view: window });
   Object.keys(options).forEach(key => event[key] = options[key]);
   let target;
   if (typeof selectorOrDomElement === 'string') {
@@ -22,16 +22,12 @@ export function nativeMouseDown(selectorOrDomElement, options = {}) {
   Ember.run(() => target.dispatchEvent(event));
 }
 
-export function nativeMouseUp(selectorOrDomElement, options = {}) {
-  let event = new window.Event('mouseup', { bubbles: true, cancelable: true, view: window });
-  Object.keys(options).forEach(key => event[key] = options[key]);
-  let target;
-  if (typeof selectorOrDomElement === 'string') {
-    target = Ember.$(selectorOrDomElement)[0];
-  } else {
-    target = selectorOrDomElement;
-  }
-  Ember.run(() => target.dispatchEvent(event));
+export function nativeMouseDown(selectorOrDomElement, options) {
+  fireNativeMouseEvent('mousedown', selectorOrDomElement, options);
+}
+
+export function nativeMouseUp(selectorOrDomElement, options) {
+  fireNativeMouseEvent('mouseup', selectorOrDomElement, options);
 }
 
 export function triggerKeydown(domElement, k) {
@@ -113,5 +109,25 @@ export default function() {
       }
     }
 
+  });
+
+  Ember.Test.registerAsyncHelper('removeMultipleOption', function(app, cssPath, value) {
+    const elem = find(`${cssPath} .ember-power-select-multiple-options > li:contains(${value}) > .ember-power-select-multiple-remove-btn`)[0];
+    try {
+      nativeMouseDown(elem);
+    } catch (e) {
+      console.warn('css path to remove btn not found');
+      throw e;
+    }
+  });
+
+  Ember.Test.registerAsyncHelper('clearSelected', function(app, cssPath, value) {
+    const elem = find(`${cssPath} .ember-power-select-clear-btn`)[0];
+    try {
+      nativeMouseDown(elem);
+    } catch (e) {
+      console.warn('css path to clear btn not found');
+      throw e;
+    }
   });
 }